+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
+-   Documentation: registerBeforeInsertTransform and registerBeforeUpdateTransform.
 -   Allow Date columns in `findByPrimaryKey`.
 -   Tested with Node.js v14.11.0.
--   Documentation: registerBeforeInsertTransform and registerBeforeUpdateTransform.
+-   Added "Breaking changes in next major release" in README.
 
 ## 2.15.0