+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 2.16.0
+
 -   Documentation: registerBeforeInsertTransform and registerBeforeUpdateTransform.
 -   Allow Date columns in `findByPrimaryKey`.
 -   Tested with Node.js v14.11.0.